Changed the cmd_in_path test to not go through SSH. We're installing the openvdc-cli package directly in our docker container
